split pg declarations
