+++ setup.py
@@ -27,7 +27,8 @@ else:
   print("Debug build")
   plugin_compile_args.extend(["-g", "-O0"])
 
-plugin_sources      = ["src/torch_ucc.cpp"]
+plugin_sources      = ["src/torch_ucc.cpp",
+                       "src/torch_ucc_comm.cpp"]
 plugin_include_dirs = ["{}/include/".format(ucc_plugin_dir),
                        "{}/include/".format(ucx_home),
                        "{}/include/".format(ucc_home)]
