+++ src/torch_ucc.cpp
@@ -140,53 +140,55 @@ void ProcessGroupUCC::WorkUCC::finalize() {
   }
 }
 
-CommUCX::CommUCX(int comm_size) {
-  ucp_params_t params;
-  ucp_config_t* config;
-  ucs_status_t st;
-  ucp_worker_params_t worker_params;
-
-  st = ucp_config_read("TORCH", nullptr, &config);
-  if (st != UCS_OK) {
-    LOG(ERROR) << "failed to read UCP config: " << ucs_status_string(st);
-    throw std::runtime_error(ucs_status_string(st));
+CommPG::CommPG(torch_ucc_oob_coll_info_t* oob_info,
+    c10::Device dev)
+    : ucx_comm(oob_info->size),
+      ucc_comm(oob_info),
+      cuda_device_index(TORCH_UCC_DEVICE_NOT_SET) {
+  if (dev.is_cuda()) {
+    cuda_device_index = dev.index();
   }
-  memset(&params, 0, sizeof(ucp_params_t));
-  params.field_mask = UCP_PARAM_FIELD_FEATURES | UCP_PARAM_FIELD_REQUEST_SIZE |
-      UCP_PARAM_FIELD_ESTIMATED_NUM_EPS | UCP_PARAM_FIELD_TAG_SENDER_MASK |
-      UCP_PARAM_FIELD_REQUEST_INIT | UCP_PARAM_FIELD_REQUEST_CLEANUP;
-  params.request_size = sizeof(ucc_coll_req_t);
-  params.features = UCP_FEATURE_TAG;
-  params.estimated_num_eps = comm_size;
-  params.tag_sender_mask = TORCH_UCX_RANK_MASK;
-  params.request_init = [](void* request) {
-    static_cast<ucc_coll_req_h>(request)->status = UCC_INPROGRESS;
-  };
-  params.request_cleanup = [](void*) {};
-  st = ucp_init(&params, config, &context);
-  ucp_config_release(config);
-  if (st != UCS_OK) {
-    LOG(ERROR) << "failed to init UCP context: " << ucs_status_string(st);
-    throw std::runtime_error(ucs_status_string(st));
-  }
-  memset(&worker_params, 0, sizeof(ucp_worker_params_t));
-  worker_params.field_mask = UCP_WORKER_PARAM_FIELD_THREAD_MODE;
-  worker_params.thread_mode = UCS_THREAD_MODE_MULTI;
-  st = ucp_worker_create(context, &worker_params, &worker);
-  if (st != UCS_OK) {
-    LOG(ERROR) << "failed to create UCP worker: " << ucs_status_string(st);
-    ucp_cleanup(context);
-    throw std::runtime_error(ucs_status_string(st));
-  }
+  stop_progress_loop = false;
+  progress_thread = std::thread(&CommPG::progress_loop, this);
+  pthread_setname_np(progress_thread.native_handle(), "ucc-progress");
 }
 
-void CommUCX::progress() {
-  ucp_worker_progress(worker);
+CommPG::~CommPG() {
+  std::unique_lock<std::mutex> lock(mutex);
+  queue_consume_cv.wait(lock, [&] { return progress_queue.empty(); });
+  stop_progress_loop = true;
+  lock.unlock();
+  queue_produce_cv.notify_all();
+  progress_thread.join();
 }
 
-CommUCX::~CommUCX() {
-  ucp_worker_destroy(worker);
-  ucp_cleanup(context);
+std::shared_ptr<CommPG> CommPG::get_comm(
+    uint32_t& id,
+    c10::Device dev,
+    torch_ucc_oob_coll_info_t *oob) {
+  static std::mutex m;
+  static std::weak_ptr<CommPG> comm;
+  static uint32_t comm_id;
+
+  std::lock_guard<std::mutex> lock(m);
+  id = (comm_id++ % TORCH_UCX_COMM_BITS);
+  std::shared_ptr<CommPG> shared_comm = comm.lock();
+  if (!shared_comm) {
+    shared_comm = std::make_shared<CommPG>(oob, dev);
+    comm = shared_comm;
+  } else {
+    if (dev.is_cuda()) {
+      if ((shared_comm->cuda_device_index != TORCH_UCC_DEVICE_NOT_SET) &&
+          (shared_comm->cuda_device_index != dev.index())) {
+        LOG(ERROR)
+            << "ucc communicator was initialized with different cuda device,"
+            << "multi device is not supported";
+        throw std::runtime_error(ucc_status_string(UCC_ERR_NOT_SUPPORTED));
+      }
+      shared_comm->cuda_device_index = dev.index();
+    }
+  }
+  return shared_comm;
 }
 
 void CommPG::ucx_connect_eps(
@@ -299,130 +301,6 @@ ucc_coll_req_h CommPG::recv_nb(
   return reinterpret_cast<ucc_coll_req_h>(st);
 }
 
-static ucc_status_t oob_allgather(
-    void* sbuf,
-    void* rbuf,
-    size_t msglen,
-    void* coll_info,
-    void** req) {
-  torch_ucc_oob_coll_info_t* info =
-      reinterpret_cast<torch_ucc_oob_coll_info_t*>(coll_info);
-  std::vector<uint8_t> val = std::vector<uint8_t>(
-      reinterpret_cast<uint8_t*>(sbuf),
-      reinterpret_cast<uint8_t*>(sbuf) + msglen);
-  info->store->set("teamr" + std::to_string(info->rank), val);
-  info->rbuf = rbuf;
-  info->msglen = msglen;
-  *req = coll_info;
-  return UCC_OK;
-}
-
-static ucc_status_t oob_allgather_test(void* req) {
-  torch_ucc_oob_coll_info_t* info =
-      reinterpret_cast<torch_ucc_oob_coll_info_t*>(req);
-
-  for (int r = 0; r < info->size; r++) {
-    if (!(info->store->check({"teamr" + std::to_string(r)}))) {
-      return UCC_INPROGRESS;
-    }
-  }
-  for (int r = 0; r < info->size; r++) {
-    std::vector<uint8_t> data =
-        info->store->get("teamr" + std::to_string(r));
-    memcpy(
-        (void*)((ptrdiff_t)info->rbuf + info->msglen * r),
-        data.data(),
-        info->msglen);
-  }
-  return UCC_OK;
-}
-
-static ucc_status_t oob_allgather_free(void* req) {
-  torch_ucc_oob_coll_info_t* info =
-      reinterpret_cast<torch_ucc_oob_coll_info_t*>(req);
-  int num_done = info->store->add({"team_ag_done"}, 1);
-  if (num_done == info->size) {
-    info->store->deleteKey("team_ag_done");
-    for (int r = 0; r < info->size; r++) {
-      info->store->deleteKey("teamr" + std::to_string(r));
-    }
-    for (int r = 0; r < info->size; r++) {
-      info->store->add({"team_ag_finished" + std::to_string(r)}, 1);
-    }
-  } else {
-    info->store->wait({"team_ag_finished" + std::to_string(info->rank)});
-  }
-  info->store->deleteKey("team_ag_finished" + std::to_string(info->rank));
-
-  return UCC_OK;
-}
-
-CommUCC::CommUCC(torch_ucc_oob_coll_info_t* oob_info) {
-  ucc_lib_config_h lib_config;
-  ucc_context_config_h context_config;
-  ucc_lib_params_t lib_params;
-  ucc_context_params_t context_params;
-  ucc_status_t st;
-
-  st = ucc_lib_config_read("TORCH", nullptr, &lib_config);
-  if (st != UCC_OK) {
-    LOG(ERROR) << "failed to read UCC lib config: " << ucc_status_string(st);
-    throw std::runtime_error(ucc_status_string(st));
-  }
-  memset(&lib_params, 0, sizeof(ucc_lib_params_t));
-  lib_params.mask = UCC_LIB_PARAM_FIELD_THREAD_MODE;
-  lib_params.thread_mode = UCC_THREAD_MULTIPLE;
-  st = ucc_init(&lib_params, lib_config, &lib);
-  ucc_lib_config_release(lib_config);
-  if (st != UCC_OK) {
-    LOG(ERROR) << "failed to init UCC lib: " << ucc_status_string(st);
-    throw std::runtime_error(ucc_status_string(st));
-  }
-  st = ucc_context_config_read(lib, NULL, &context_config);
-  if (st != UCC_OK) {
-    ucc_finalize(lib);
-    LOG(ERROR) << "failed to read UCC context config: "
-               << ucc_status_string(st);
-    throw std::runtime_error(ucc_status_string(st));
-  }
-  st = ucc_context_config_modify(
-      context_config,
-      NULL,
-      "ESTIMATED_NUM_EPS",
-      std::to_string(oob_info->size).c_str());
-  if (st != UCC_OK) {
-    ucc_context_config_release(context_config);
-    ucc_finalize(lib);
-    LOG(ERROR) << "failed to modify UCC context config: "
-               << ucc_status_string(st);
-    throw std::runtime_error(ucc_status_string(st));
-  }
-  memset(&context_params, 0, sizeof(ucc_context_params_t));
-  context_params.mask =
-      UCC_CONTEXT_PARAM_FIELD_TYPE | UCC_CONTEXT_PARAM_FIELD_OOB;
-  context_params.type = UCC_CONTEXT_SHARED;
-  context_params.oob.allgather = oob_allgather;
-  context_params.oob.req_test = oob_allgather_test;
-  context_params.oob.req_free = oob_allgather_free;
-  context_params.oob.coll_info = oob_info;
-  ucc_context_create(lib, &context_params, context_config, &context);
-  ucc_context_config_release(context_config);
-  if (st != UCC_OK) {
-    ucc_finalize(lib);
-    LOG(ERROR) << "failed to create UCC context: " << ucc_status_string(st);
-    throw std::runtime_error(ucc_status_string(st));
-  }
-}
-
-void CommUCC::progress() {
-  ucc_context_progress(context);
-}
-
-CommUCC::~CommUCC() {
-  ucc_context_destroy(context);
-  ucc_finalize(lib);
-}
-
 void CommPG::ucc_create_team(
     ucc_team_h& team,
     torch_ucc_oob_coll_info_t* oob_info) {
@@ -454,6 +332,121 @@ void CommPG::ucc_create_team(
 void CommPG::ucc_destroy_team(ucc_team_h& team) {
   ucc_team_destroy(team);
 }
+
+c10::intrusive_ptr<ProcessGroup::Work> CommPG::enqueue_p2p(
+    OpType opType,
+    ucc_coll_req_h request) {
+  if (request == nullptr) {
+    // p2p2 request completed immediately don't save it to progress queue
+    return c10::make_intrusive<ProcessGroupUCC::WorkUCC>(
+        opType, UCC_OK, request, nullptr, &ucx_comm);
+  }
+  std::unique_lock<std::mutex> lock(mutex);
+  auto work = c10::make_intrusive<ProcessGroupUCC::WorkUCC>(
+      opType, UCC_INPROGRESS, request, nullptr, &ucx_comm);
+  progress_queue.push_back(work);
+  lock.unlock();
+  queue_produce_cv.notify_one();
+  return work;
+}
+
+c10::intrusive_ptr<ProcessGroupUCC::WorkUCC> CommPG::enqueue_collective(
+    OpType opType,
+    ucc_coll_args_t& coll,
+    std::unique_ptr<ProcessGroupUCC::WorkData> data,
+    ucc_team_h& team) {
+  std::unique_lock<std::mutex> lock(mutex);
+  ucc_coll_req_h request;
+  ucc_status_t st;
+  st = ucc_collective_init(&coll, &request, team);
+  if (st != UCC_OK) {
+    LOG(ERROR) << "failed to init collective: " << ucc_status_string(st);
+    throw std::runtime_error(ucc_status_string(st));
+  }
+  st = ucc_collective_post(request);
+  if (st != UCC_OK) {
+    LOG(ERROR) << "failed to post collective: " << ucc_status_string(st);
+    throw std::runtime_error(ucc_status_string(st));
+  }
+  auto work = c10::make_intrusive<ProcessGroupUCC::WorkUCC>(
+      opType, UCC_INPROGRESS, request, nullptr, &ucc_comm);
+  work->data = std::move(data);
+  progress_queue.push_back(work);
+  lock.unlock();
+  queue_produce_cv.notify_one();
+  return work;
+}
+
+#ifdef USE_CUDA
+c10::intrusive_ptr<ProcessGroupUCC::WorkUCC> CommPG::enqueue_cuda_collective(
+    OpType opType,
+    ucc_coll_args_t& coll,
+    std::unique_ptr<ProcessGroupUCC::WorkData> data,
+    ucc_team_h& team,
+    ucc_ee_h ee) {
+  std::unique_lock<std::mutex> lock(mutex);
+  ucc_coll_req_h request;
+  ucc_status_t st;
+  st = ucc_collective_init(&coll, &request, team);
+  if (st != UCC_OK) {
+    LOG(ERROR) << "failed to init collective: " << ucc_status_string(st);
+    throw std::runtime_error(ucc_status_string(st));
+  }
+  ucc_ev_t comp_ev, *post_ev;
+  comp_ev.ev_type = UCC_EVENT_COMPUTE_COMPLETE;
+  comp_ev.ev_context = nullptr;
+  comp_ev.ev_context_size = 0;
+  comp_ev.req = request;
+  st = ucc_collective_triggered_post(ee, &comp_ev);
+  if (st != UCC_OK) {
+    LOG(ERROR) << "failed to post triggered collective: "
+                << ucc_status_string(st);
+    throw std::runtime_error(ucc_status_string(st));
+  }
+  st = ucc_ee_get_event(ee, &post_ev);
+  TORCH_CHECK(st == UCC_OK && post_ev->ev_type == UCC_EVENT_COLLECTIVE_POST);
+  ucc_ee_ack_event(ee, post_ev);
+  auto work = c10::make_intrusive<ProcessGroupUCC::WorkUCC>(
+      opType, UCC_INPROGRESS, request, ee, &ucc_comm);
+  work->data = std::move(data);
+  progress_queue.push_back(work);
+  lock.unlock();
+  queue_produce_cv.notify_one();
+  return work;
+}
+#endif
+
+void CommPG::progress_loop() {
+  std::unique_lock<std::mutex> lock(mutex);
+#ifdef USE_CUDA
+  bool device_set = false;
+#endif
+  while (!stop_progress_loop) {
+    if (progress_queue.empty()) {
+      queue_produce_cv.wait(lock);
+      continue;
+    }
+    auto work = progress_queue.front();
+    progress_queue.pop_front();
+    lock.unlock();
+    queue_consume_cv.notify_one();
+#ifdef USE_CUDA
+    if ((!device_set) && (cuda_device_index != TORCH_UCC_DEVICE_NOT_SET)) {
+      c10::cuda::set_device(cuda_device_index);
+      device_set = true;
+    }
+#endif
+    while (work->request_->status > 0) {
+      // operation initialized is in progress or
+      work->comm_->progress();
+    }
+
+    lock.lock();
+    work->finalize();
+  }
+}
+
+
 
 ProcessGroupUCC::ProcessGroupUCC(
     const c10::intrusive_ptr<Store>& store,
